✒️Feat: 컴포넌트 폴더 변경
