+++ packages/parang/src/components/index.js
@@ -1,3 +1,5 @@
-export * from './bookmark.js';
-export * from './explore.js';
-export * from './profile.js';
+export * from './login/auth';
+
+export * from './main/bookmark';
+export * from './main/explore';
+export * from './main/profile';
