+++ packages/parang/src/components/main/bookmark.js
@@ -1,0 +1,3 @@
+export class Bookmark {
+  constructor() {}
+}
