+++ packages/parang/src/components/main/profile.js
@@ -1,0 +1,3 @@
+export class Profile {
+  constructor() {}
+}
